[Core] Move MRV1 late_interaction_runner.py out of MRV2 subtree
This file was added in https://github.com/vllm-project/vllm/pull/36159 but inadvertently put into the MRV2 source subtree.
